Heavy refactoring.
Changed socket.hpp parts not to rely at all on std::unique_ptr.
instead the std::optional is used to allow for looping until
connections have been received. Using move-semantics and allowing
the move operations allows the caller to decide if he wants to
use unique_ptr or whatever for storing the values.

Removed storing the source and destination in the class of
sockets. They can be queried from the OS. Especially since
they're mostly used for debug functionality.